BHIVE-206 fix global key error

+++ lib/features/home/view/widgets/home_widget.dart
@@ -48,7 +48,7 @@ class HomeBanner extends StatelessWidget {
         ),
         //dots
         DotsIndicator(
-          position: ref.watch(homeScreenBannerDotsProvider),
+          position: ref.watch(homeScreenBannerDotsProvider).toDouble(),
           dotsCount: 3,
           mainAxisAlignment: MainAxisAlignment.center,
           decorator: DotsDecorator(
